refactor(gimp): extract frontend components
Extract ChatMessage, StatusBar, and DevTools into
src/lib/components/ and update +page.svelte to use them.
Reduces page from ~720 lines to ~330 lines.
