+++ apps/gimp-assistant/src/routes/+page.svelte
@@ -1,6 +1,9 @@
 <script lang="ts">
   import { invoke, Channel } from "@tauri-apps/api/core";
   import { onMount } from "svelte";
+  import ChatMessage from "$lib/components/ChatMessage.svelte";
+  import StatusBar from "$lib/components/StatusBar.svelte";
+  import DevTools from "$lib/components/DevTools.svelte";
 
   type AssistantResponse = {
     reply: string;
@@ -36,17 +39,6 @@
   let showDevTools = $state(false);
   let chatEl = $state<HTMLElement | undefined>(undefined);
   let textareaEl = $state<HTMLTextAreaElement | undefined>(undefined);
-
-  // Dev tools state
-  let engineTestStatus = $state("Unknown");
-  let gimpStatus = $state("Unknown");
-  let engineTestResult = $state("");
-  let toolsListResult = $state("");
-  let actionLog = $state<string[]>([]);
-
-  function logAction(msg: string) {
-    actionLog = [msg, ...actionLog].slice(0, 20);
-  }
 
   // Auto-scroll chat to bottom whenever messages update
   $effect(() => {
@@ -105,7 +97,6 @@
     messages = [...messages, { role: "user", text: trimmed }];
     isSending = true;
 
-    // Add a streaming placeholder message
     const streamIdx = messages.length;
     messages = [...messages, { role: "assistant", text: "", isStreaming: true }];
 
@@ -115,7 +106,6 @@
     channel.onmessage = (token: string) => {
       accumulated += token;
       isStreaming = true;
-      // Update the streaming message in-place
       messages = messages.map((m, i) =>
         i === streamIdx ? { ...m, text: accumulated } : m
       );
@@ -127,7 +117,6 @@
         onToken: channel,
       });
 
-      // Finalize the message: use accumulated text if streamed, otherwise use reply
       const finalText = result.streamed
         ? accumulated
         : (result.reply || accumulated || "Done.");
@@ -193,57 +182,6 @@
     if (!textareaEl) return;
     textareaEl.style.height = "auto";
   }
-
-  // Dev tools helpers
-  async function testEngine() {
-    engineTestStatus = "Checking\u2026";
-    try {
-      const healthy = await invoke<boolean>("engine_health");
-      engineTestResult = healthy ? "Engine is healthy" : "Engine not ready";
-      engineTestStatus = healthy ? "Connected" : "Offline";
-      engineStatus = healthy ? "ready" : "offline";
-    } catch (e) {
-      engineTestResult = String(e);
-      engineTestStatus = "Error";
-      engineStatus = "offline";
-    }
-  }
-
-  async function listTools() {
-    gimpStatus = "Checking\u2026";
-    try {
-      const result = await invoke<any>("mcp_list_tools");
-      toolsListResult = JSON.stringify(result, null, 2);
-      gimpStatus = "Connected";
-    } catch (e) {
-      toolsListResult = String(e);
-      gimpStatus = "Disconnected";
-    }
-  }
-
-  async function runDrawTestLine() {
-    logAction("Draw test line\u2026");
-    try {
-      await invoke("macro_draw_line", { x1: 50, y1: 50, x2: 200, y2: 200 });
-      logAction("Line OK");
-    } catch (e) { logAction("Failed: " + String(e)); }
-  }
-
-  async function runCropSquare() {
-    logAction("Crop square\u2026");
-    try {
-      await invoke("macro_crop_square");
-      logAction("Crop OK");
-    } catch (e) { logAction("Failed: " + String(e)); }
-  }
-
-  async function runResize1024() {
-    logAction("Resize to 1024w\u2026");
-    try {
-      await invoke("macro_resize", { width: 1024 });
-      logAction("Resize OK");
-    } catch (e) { logAction("Failed: " + String(e)); }
-  }
 </script>
 
 <div class="app">
@@ -257,18 +195,7 @@
       </div>
     </div>
     <div class="header-right">
-      <div class="status-pill" class:connected={engineStatus === "ready"} class:offline={engineStatus === "offline"}>
-        <span class="status-dot"></span>
-        <span class="status-label">
-          {engineStatus === "ready" ? "Engine ready" : engineStatus === "offline" ? "Engine offline" : "Engine..."}
-        </span>
-      </div>
-      <div class="status-pill" class:connected={isConnected}>
-        <span class="status-dot"></span>
-        <span class="status-label">
-          {isConnected ? (imageInfo || "GIMP connected") : "GIMP offline"}
-        </span>
-      </div>
+      <StatusBar {engineStatus} {isConnected} {imageInfo} />
       <button
         class="icon-btn"
         onclick={() => (showDevTools = !showDevTools)}
@@ -299,28 +226,10 @@
 
       <!-- Messages -->
       {#each messages as msg}
-        <div class="row {msg.role}">
-          {#if msg.role === "assistant"}
-            <div class="avatar">✦</div>
-          {/if}
-          <div class="col">
-            <div class="bubble {msg.role}">
-              {msg.text}{#if msg.isStreaming}<span class="cursor">|</span>{/if}
-            </div>
-            {#if msg.explain}
-              <div class="tip">
-                <span class="tip-icon">💡</span>
-                <span class="tip-text">{msg.explain}</span>
-              </div>
-            {/if}
-            {#if msg.undoable}
-              <button class="undo-btn" onclick={undoLast}>↩ Undo</button>
-            {/if}
-          </div>
-        </div>
+        <ChatMessage {msg} onUndo={undoLast} />
       {/each}
 
-      <!-- Typing indicator (only when sending but not yet streaming) -->
+      <!-- Typing indicator -->
       {#if isSending && !isStreaming && messages[messages.length - 1]?.text === ""}
         <div class="row assistant">
           <div class="avatar">✦</div>
@@ -363,36 +272,7 @@
 
   <!-- Dev tools overlay -->
   {#if showDevTools}
-    <aside class="devtools">
-      <div class="devtools-header">
-        <span>Developer Tools</span>
-        <button class="icon-btn" onclick={() => (showDevTools = false)}>✕</button>
-      </div>
-
-      <details class="dev-section">
-        <summary>Engine · <em>{engineTestStatus}</em></summary>
-        <button class="dev-btn" onclick={testEngine}>Test Engine</button>
-        {#if engineTestResult}<pre>{engineTestResult}</pre>{/if}
-      </details>
-
-      <details class="dev-section">
-        <summary>GIMP MCP · <em>{gimpStatus}</em></summary>
-        <button class="dev-btn" onclick={listTools}>Refresh Tools</button>
-        {#if toolsListResult}<pre>{toolsListResult}</pre>{/if}
-      </details>
-
-      <details class="dev-section" open>
-        <summary>Quick Actions</summary>
-        <div class="dev-actions">
-          <button class="dev-btn" onclick={runDrawTestLine}>Line</button>
-          <button class="dev-btn" onclick={runCropSquare}>Crop</button>
-          <button class="dev-btn" onclick={runResize1024}>1024w</button>
-        </div>
-        {#if actionLog.length > 0}
-          <pre class="action-log">{actionLog.join("\n")}</pre>
-        {/if}
-      </details>
-    </aside>
+    <DevTools onClose={() => (showDevTools = false)} />
   {/if}
 </div>
 
@@ -454,41 +334,6 @@
     align-items: center;
     gap: 8px;
   }
-
-  /* Status pill */
-  .status-pill {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    padding: 4px 10px;
-    border-radius: 20px;
-    background: #f5f5f5;
-    border: 1px solid #e0e0e0;
-    font-size: 12px;
-    color: #888;
-    transition: background 0.2s, color 0.2s;
-  }
-  .status-pill.connected {
-    background: #f0faf3;
-    border-color: #b8e6c6;
-    color: #2a7a45;
-  }
-  .status-pill.offline {
-    background: #fef2f2;
-    border-color: #fecaca;
-    color: #b91c1c;
-  }
-  .status-dot {
-    width: 7px;
-    height: 7px;
-    border-radius: 50%;
-    background: #ccc;
-    flex-shrink: 0;
-    transition: background 0.2s;
-  }
-  .status-pill.connected .status-dot { background: #34c759; }
-  .status-pill.offline .status-dot { background: #ef4444; }
-  .status-label { white-space: nowrap; max-width: 140px; overflow: hidden; text-overflow: ellipsis; }
 
   /* Icon button */
   .icon-btn {
@@ -580,15 +425,13 @@
     color: #fff;
   }
 
-  /* Message rows */
+  /* Typing indicator (kept in page since it's not a reusable message) */
   .row {
     display: flex;
     align-items: flex-end;
     gap: 8px;
     max-width: 100%;
   }
-  .row.user { flex-direction: row-reverse; }
-
   .avatar {
     width: 28px;
     height: 28px;
@@ -602,7 +445,6 @@
     flex-shrink: 0;
     margin-bottom: 2px;
   }
-
   .col {
     display: flex;
     flex-direction: column;
@@ -610,9 +452,6 @@
     max-width: 72%;
     gap: 6px;
   }
-  .row.user .col { align-items: flex-end; }
-
-  /* Bubbles */
   .bubble {
     padding: 10px 14px;
     border-radius: 18px;
@@ -621,29 +460,11 @@
     word-break: break-word;
     white-space: pre-wrap;
   }
-  .bubble.user {
-    background: #007aff;
-    color: #fff;
-    border-bottom-right-radius: 5px;
-  }
   .bubble.assistant {
     background: #f0f0f5;
     color: #1a1a1a;
     border-bottom-left-radius: 5px;
   }
-
-  /* Streaming cursor */
-  .cursor {
-    animation: blink 0.8s step-end infinite;
-    font-weight: 300;
-    color: #007aff;
-  }
-  @keyframes blink {
-    0%, 100% { opacity: 1; }
-    50% { opacity: 0; }
-  }
-
-  /* Typing indicator */
   .bubble.typing {
     display: flex;
     align-items: center;
@@ -664,33 +485,6 @@
     0%, 60%, 100% { transform: translateY(0); opacity: 0.5; }
     30% { transform: translateY(-5px); opacity: 1; }
   }
-
-  /* Explain tip */
-  .tip {
-    display: flex;
-    gap: 8px;
-    align-items: flex-start;
-    background: #fffbea;
-    border-left: 3px solid #f5c842;
-    border-radius: 8px;
-    padding: 8px 12px;
-    max-width: 100%;
-  }
-  .tip-icon { font-size: 13px; flex-shrink: 0; margin-top: 2px; }
-  .tip-text { font-size: 12px; color: #5a4a00; line-height: 1.5; margin: 0; }
-
-  /* Undo button */
-  .undo-btn {
-    font-size: 12px;
-    padding: 4px 12px;
-    border-radius: 12px;
-    border: 1px solid rgba(0, 122, 255, 0.3);
-    background: rgba(0, 122, 255, 0.06);
-    color: #007aff;
-    cursor: pointer;
-    transition: background 0.15s;
-  }
-  .undo-btn:hover { background: rgba(0, 122, 255, 0.14); }
 
   /* Input bar */
   .input-bar {
@@ -755,72 +549,4 @@
     line-height: 1;
   }
   .stop-btn:hover { background: #dc2626; }
-
-  /* Dev tools overlay */
-  .devtools {
-    position: absolute;
-    top: 57px;
-    right: 0;
-    bottom: 0;
-    width: 280px;
-    background: #fff;
-    border-left: 1px solid #e0e0e8;
-    display: flex;
-    flex-direction: column;
-    overflow-y: auto;
-    z-index: 10;
-    box-shadow: -4px 0 16px rgba(0,0,0,0.06);
-  }
-  .devtools-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 12px 14px;
-    border-bottom: 1px solid #eee;
-    font-size: 13px;
-    font-weight: 600;
-    color: #333;
-    flex-shrink: 0;
-  }
-  .dev-section {
-    border-bottom: 1px solid #f0f0f0;
-    font-size: 12px;
-  }
-  .dev-section > summary {
-    padding: 10px 14px;
-    cursor: pointer;
-    user-select: none;
-    color: #444;
-    font-weight: 500;
-    list-style: none;
-  }
-  .dev-section > summary::-webkit-details-marker { display: none; }
-  .dev-section > summary::before { content: "▶ "; font-size: 9px; }
-  .dev-section[open] > summary::before { content: "▼ "; }
-  .dev-section em { font-style: normal; color: #888; font-weight: 400; }
-  .dev-section > :not(summary) { padding: 0 14px 10px; }
-  .dev-btn {
-    margin: 4px 0;
-    padding: 5px 10px;
-    font-size: 12px;
-    border-radius: 6px;
-    border: 1px solid #d0d0d8;
-    background: #fafafa;
-    cursor: pointer;
-    transition: background 0.1s;
-  }
-  .dev-btn:hover { background: #f0f0f0; }
-  .dev-actions { display: flex; gap: 6px; flex-wrap: wrap; padding-top: 4px; }
-  pre {
-    font-size: 10px;
-    background: #f4f4f8;
-    border-radius: 6px;
-    padding: 8px;
-    overflow-x: auto;
-    white-space: pre-wrap;
-    word-break: break-all;
-    margin: 6px 0 0;
-    color: #333;
-  }
-  .action-log { margin: 8px 0 0; }
 </style>
